+++ src/vector2D.js
@@ -162,6 +162,30 @@ TYPE6JS.Vector2D = {
     this.y = 0.0;
     //return this.toString() ;
   },
+  
+  /**
+  * Set to Origin
+  * @since 0.2.1
+  * @method
+  * @param {float} [x = 0.0] x The new value of the x coordinate.
+  * @param {float} [y = 0.0] y The new value of the y coordinate.
+  */
+  setXToOrigin: function() {
+    this.x = 0.0;
+    //return this.toString() ;
+  },
+  
+  /**
+  * Set to Origin
+  * @since 0.2.1
+  * @method
+  * @param {float} [x = 0.0] x The new value of the x coordinate.
+  * @param {float} [y = 0.0] y The new value of the y coordinate.
+  */
+  setYToOrigin: function() {
+    this.y = 0.0;
+    //return this.toString() ;
+  },
 
   /**
   * set X and Y coordinates from the given angle.
@@ -261,13 +285,12 @@ TYPE6JS.Vector2D = {
   * copy the given vector2D values.
   * @since 0.2.0
   * @method
-  * @param {vector2D} vector2D The vector2D you need to copy.
-  * @returns {Vector2} A new vector2D identical to the vector2D given in paramters
+  * @returns {Vector2} A new vector2D identical to this vector2D
   */
-  copy: function(vector2D) {
+  copy: function() {
     return this.create(
-      vector2D.getX(),
-      vector2D.getY()
+      this.getX(),
+      this.getY()
     );
   },
 
@@ -332,8 +355,8 @@ TYPE6JS.Vector2D = {
   * @returns {Vector2} A new vector with the scaled vector added
   */
   addScaledVector:function(vector2D, scalar){
-    return this.create( this.x + vector2D.x * scalar,
-                        this.y + vector2D.y * scalar
+    return this.create( this.x + vector2D.getX() * scalar,
+                        this.y + vector2D.getY() * scalar
                       );
   },
 
@@ -364,6 +387,20 @@ TYPE6JS.Vector2D = {
   },
 
   /**
+  * scale the given vector by the given scalar, subtract it to this vector and creates a new vector.
+  * @since 0.0.3
+  * @method
+  * @param {Vector2} vector2D The vector to scale.
+  * @param {float} scalar The scale amount.
+  * @returns {Vector2} A new vector with the scaled vector subtracted
+  */
+  subtractScaledVector:function(vector2D, scalar){
+    return this.create( this.x - vector2D.getX() * scalar,
+                        this.y - vector2D.getY() * scalar
+                      );
+  },
+
+  /**
   * scale this vector by the given scalar and create a new vector
   * @since 0.0.3
   * @method
@@ -385,8 +422,8 @@ TYPE6JS.Vector2D = {
   */
   //component product
   multiply:function(vector2D){
-    return this.create( this.x * vector2D.x,
-                        this.y * vector2D.y
+    return this.create( this.x * vector2D.getX(),
+                        this.y * vector2D.getY()
                       );
   },
 
@@ -398,8 +435,8 @@ TYPE6JS.Vector2D = {
   * @returns {Vector2} a new vector with the new coordinates
   */
   divide: function(vector2D) {
-    return this.create( this.x / vector2D.x,
-                        this.y / vector2D.y
+    return this.create( this.x / vector2D.getX(),
+                        this.y / vector2D.getY()
                       );
   },
 
@@ -489,12 +526,12 @@ TYPE6JS.Vector2D = {
     return this.create( TYPE6JS.MathUtils.clamp(
                           this.x,
                           rectangle.topLeftCorner.getX(),
-                          rectangle.topLeftCorner.getX() + rectangle.size.getX()
+                          rectangle.bottomRightCorner.getX()
                         ),
                         TYPE6JS.MathUtils.clamp(
                           this.y,
                           rectangle.topLeftCorner.getY(),
-                          rectangle.topLeftCorner.getY() + rectangle.size.getY()
+                          rectangle.bottomRightCorner.getY()
                         )
                       );
   },
@@ -651,8 +688,8 @@ TYPE6JS.Vector2D = {
   * @param {value} value The scale amount.
   */
   addScaledVectorTo:function(vector2D, scalar){
-    this.x += vector2D.x * scalar;
-    this.y += vector2D.y * scalar;
+    this.x += vector2D.getX() * scalar;
+    this.y += vector2D.getY() * scalar;
   },
   
   /**
@@ -663,8 +700,8 @@ TYPE6JS.Vector2D = {
   * @param {value} value The the scale amount.
   */
   copyScaledVectorTo:function(vector2D, scalar){
-    this.x = vector2D.x * scalar;
-    this.y = vector2D.y * scalar;
+    this.x = vector2D.getX() * scalar;
+    this.y = vector2D.getY() * scalar;
   },
 
   /**
@@ -701,6 +738,18 @@ TYPE6JS.Vector2D = {
     this.x -= scalar;
     this.y -= scalar;
   },
+  
+  /**
+  * Subtract scaled vector to.
+  * @since 0.0.1
+  * @method
+  * @param {Vector2} value The vector to scale.
+  * @param {value} value The scale amount.
+  */
+  subtractScaledVectorFrom:function(vector2D, scalar){
+    this.x -= vector2D.getX() * scalar;
+    this.y -= vector2D.getY() * scalar;
+  },
 
   /**
   * Scale by
@@ -721,8 +770,8 @@ TYPE6JS.Vector2D = {
   */
   //component product
   multiplyBy: function(vector2D){
-    this.x *= vector2D.x;
-    this.y *= vector2D.y;
+    this.x *= vector2D.getX();
+    this.y *= vector2D.getY();
   },
 
   /**
@@ -733,8 +782,8 @@ TYPE6JS.Vector2D = {
   */
   //Prefer scale by value inferior to 1 if possible
   divideBy: function(vector2D) {
-    this.x /= vector2D.x;
-    this.y /= vector2D.y;
+    this.x /= vector2D.getX();
+    this.y /= vector2D.getY();
   },
 
   /**
@@ -812,12 +861,12 @@ TYPE6JS.Vector2D = {
     this.x = TYPE6JS.MathUtils.clamp(
               this.x,
               rectangle.topLeftCorner.getX(),
-              rectangle.topLeftCorner.getX() + rectangle.size.getX()
+              rectangle.bottomRightCorner.getX()
             );
     this.y = TYPE6JS.MathUtils.clamp(
               this.y,
               rectangle.topLeftCorner.getY(),
-              rectangle.topLeftCorner.getY() + rectangle.size.getY()
+              rectangle.bottomRightCorner.getY()
             );
   },
 
@@ -841,7 +890,7 @@ TYPE6JS.Vector2D = {
   * @returns {float}
   */
   dotProduct: function(vector2D){ //scalar product
-    return this.x * vector2D.x + this.y * vector2D.y;
+    return this.x * vector2D.getX() + this.y * vector2D.getY();
   },
 
   /**
@@ -852,14 +901,37 @@ TYPE6JS.Vector2D = {
   */
   //true if vector is equal to (0;0)
   isOrigin: function(){
-    if( !this.x || !this.y ){
+    if( this.x === 0 || this.y === 0 )
       return true;
-    }
     return false;
   },
 
   /**
-  * Check if voector is not origin
+  * Check if the two components of the vector are greater then zero
+  * @since 0.2.1
+  * @method
+  * @returns {boolean}
+  */
+  isPositive: function(){
+    if( this.getX() > 0 && this.getY() > 0 )
+      return true;
+    return false;
+  },
+  
+  /**
+  * Check if the two components of the vector are smaller then zero
+  * @since 0.2.1
+  * @method
+  * @returns {boolean}
+  */
+  isNegative: function(){
+    if( this.getX() < 0 && this.getY() < 0 )
+      return true;
+    return false;
+  },
+
+  /**
+  * Check if vector is not origin
   * @since 0.0.5
   * @method
   * @returns {boolean}